Title: 1까지 나누기, Time: 50ms, Memory: 15MB, Status: Passed - CodeTree

+++ 241104/1까지 나누기/divide-by-1.py
@@ -3,8 +3,10 @@ i = 1
 while(True):
     if (i == 1):
         a = n
-    a = a / i
+    a = int(a / i)
+
     if (a <= 1):
         print(i)
         break
+
     i += 1
